+++ tests/testthat/test_files.R
@@ -39,11 +39,11 @@ test_that("rename_with_nums works", {
   expect_true(dir.remove("rename_with_nums_test"))
 })
 
-test_that("create_dirs works", {
+test_that("create_dir works", {
   setwd(tempdir())
-  expect_equal(create_dirs(c("mydir", "yourdir")), rep(TRUE, 2),
+  expect_equal(create_dir(c("mydir", "yourdir")), rep(TRUE, 2),
                check.names = FALSE)
-  expect_equal(create_dirs(c("mydir", "yourdir")), rep(FALSE, 2),
+  expect_equal(create_dir(c("mydir", "yourdir")), rep(FALSE, 2),
                check.names = FALSE)
   expect_equal(dir.remove(c("mydir", "yourdir")), rep(TRUE, 2),
                check.names = FALSE)
@@ -68,26 +68,3 @@ test_that("file.move errors correctly", {
   file.create("tmpfile0.R")
   expect_error(file.move("tmpfile0.R", c("tmpdir0", "tmpdir0")))
 })
-
-test_that("merge_tables_on_disk works", {
-  setwd(tempdir())
-  expect_true(dir.create("merge_tables_on_disk_test"))
-  setwd("merge_tables_on_disk_test")
-  tab1 <- tibble::tibble(x = 1.5, y = 2.5)
-  tab2 <- tibble::tibble(x = 1.5, y = 29.5)
-  tab3 <- tibble::tibble(x = 1.5, z = 29.5)
-  tab4 <- tibble::tibble(x = 1.5, y = 29.5, z = 0.5)
-  mapply(readr::write_csv, list(tab1, tab2, tab3, tab4),
-         paste0(c("tab1", "tab2", "tab3", "tab4"), ".csv"))
-  expect_equal(merge_tables_on_disk(c("tab1.csv", "tab2.csv"), ",", "merged.csv"),
-               tibble::tibble(x = c(1.5, 1.5), y = c(2.5, 29.5)),
-               check.attributes = FALSE)
-  expect_equal(readr::read_csv("merged.csv"),
-               tibble::tibble(x = c(1.5, 1.5), y = c(2.5, 29.5)),
-               check.attributes = FALSE)
-  expect_error(merge_tables_on_disk(c("tab1.csv", "tab3.csv"), ",", "merged.csv"))
-  expect_error(merge_tables_on_disk(c("tab1.csv", "tab4.csv"), ",", "merged.csv"))
-  setwd("..")
-  expect_true(dir.remove("merge_tables_on_disk_test"))
-})
-